Fixed height of 2 problem we were having with noise generation

+++ Chunk.java
@@ -63,7 +63,7 @@ public class Chunk {
         FloatBuffer VertexColorData = BufferUtils.createFloatBuffer((CHUNK_SIZE * CHUNK_SIZE * CHUNK_SIZE) * 6 * 12);
         for (float x = 0; x < CHUNK_SIZE; x+=1) {
             for (float z = 0; z < CHUNK_SIZE; z+=1) {
-                double height = (startY + (int)(14*((noise.getNoise((int)x, (int)CHUNK_SIZE, (int)z)) + 1)/2)) * CUBE_LENGTH;
+                double height = (startY + (double)(14*((noise.getNoise((int)x, (int)CHUNK_SIZE, (int)z)) + 1)/2)) * CUBE_LENGTH;
                 for (float y = 0; y < height; y++) {
                     VertexPositionData.put(createCube((float) (startX + x * CUBE_LENGTH), (float)(y * CUBE_LENGTH + (int)(CHUNK_SIZE * .8)),(float) (startZ + z *CUBE_LENGTH)));
                     VertexColorData.put(createCubeVertexCol(getCubeColor(Blocks[(int) x][(int) y][(int) z])));
